requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.7.0
arxiv==2.1.3
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.6
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.16.1
fsspec==2024.12.0
gradio==5.9.1
gradio_client==1.5.2
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.0
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.2.0
ollama==0.4.4
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==2.10.1
protobuf==5.29.2
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pywin32==308
PyYAML==6.0.2
qdrant-client==1.12.1
requests==2.32.3
rich==13.9.4
ruff==0.8.4
safehttpx==0.1.6
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
websockets==14.1
